create environment module, moved prep services and setupsettings
